Change IDs from i64 to u32

+++ src/models/auth_model.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Claims {
-    pub user_id: i64,
+    pub user_id: u32,
     pub iat: usize,
     pub exp: usize,
 }
