+++ src/schemas/city_schema.rs
@@ -6,14 +6,14 @@ use super::country_schema::CountrySummarizedSchema;
 #[derive(Debug, Serialize, TS)]
 #[ts(export)]
 pub struct CitySummarizedSchema {
-    pub id: i64,
+    pub id: u32,
     pub name: String,
 }
 
 #[derive(Debug, Serialize, TS)]
 #[ts(export)]
 pub struct CityExtendedSchema {
-    pub id: i64,
+    pub id: u32,
     pub name: String,
     pub country: CountrySummarizedSchema,
 }
